continuing the shifts fix. reviewed ok.

+++ HarmonyDB.Index/HarmonyDB.Index.Analysis.Em/Models/Constants.cs
@@ -12,7 +12,7 @@ public static class Constants
             .ToList();
 
     public static IReadOnlyList<(byte tonic, Scale scale)> Pairs { get; }
-        = Indices.Select(x => ((byte)x.tonic, (Scale)x.scale)).ToList();
+        = Indices.Select(x => (x.tonic, (Scale)x.scale)).ToList();
 
     public static int GetMajorTonic((byte tonic, Scale scale) scale, bool isSong)
     {
@@ -21,7 +21,6 @@ public static class Constants
 
     public static (byte tonic, Scale scale) GetParallelScale((byte tonic, Scale scale) scale, bool isSong)
     {
-        // todo: write unit tests
         return scale.scale == Scale.Major
             ? ((byte)((scale.tonic + (isSong ? -3 : 3) + TonicCount) % TonicCount), Scale.Minor) 
             : ((byte)((scale.tonic + (isSong ? 3 : -3) + TonicCount) % TonicCount), Scale.Major);
